lambda/requirements.txt
boto3==1.34.34
botocore==1.34.34
opensearch-py==2.4.2
requests-aws4auth==1.2.3
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.6.0
